backend/requirements.txt
# ─── FastAPI + Server ───
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# ─── Database ───
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# ─── Background Tasks ───
celery==5.3.6
redis==5.0.1

# ─── Auth ───
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# ─── Core Video Pipeline ───
google-api-python-client==2.108.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
edge-tts==7.2.7
feedparser==6.0.11
ffmpeg-python==0.2.0
pillow==10.1.0
requests==2.31.0

# ─── Config & Utils ───
python-dotenv==1.0.0
PyYAML
psutil
email-validator

# ─── Cloud AI (Gemini + Groq Fallback) ───
google-genai>=1.0.0
groq>=0.4.0

# ─── Testing ───
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
pytest-cov==4.1.0
